[BIT-138] Moved functions to new classes
